doc: continue fleshing out doc organization
Add more infrastructure organization for content published to
opea-project.github.io with some work-in-progress document and pointers
to relevent github.com documents for placeholders.  This gives us some
structure for working on a getting started guide and community content.

Add a grid to the home page for top documentation areas.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>