doc: add pymarkdown tool
pymarkdown provides checks for properly written markdown files.
It's quite configurable and checks for many items we choose not to
enforce, such as 80 character line length.  It can be configured so a CI
system could do checks on markdown files being submitted in a PR, but
for now, I'm just including the tool in the the list of packages
installed from pypi and a script to call it (checkmd.sh) with some
settings that limit what checks we're interested in (for now).

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>