doc: auto toctree for examples and microservices
Instead of manually maintaining the headings and toctree for examples
and microservices, build the toctree using a script during doc build
time that scans for directories in GenAIExamples and GenAIComps/comps.

This way, if new examples or comps show up (or are renamed) they're
automatically added to these index documents for github.io
publishing.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>