doc: fix relative images and raw HTML
Displaying images using raw HTML instead of normal markdown tags is not
processed correctly for github.io when relative links to the images are
used. Convert the multiple images into one and reference the new images
instead of multiple smaller ones, and convert all the raw HTML to
markdown references if relative links were used.  This way the content
displays well when reading the markdown file on github.com and on the
rendered opea-project.github.io site.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>