doc: fix missing references to README.md
Cross-repo references using an URL that only has the folder name cause
links in the github.io site to go to the github.com site.  Most of these
uses want to link to the documentation (README.md) so add that where
appropriate so links in the github.io site stay there.

Now that we've got those links fixed, we can reduce the complexity in
the fix-github-md-refs script that was trying to add a README.md to
folder names. Also added an area to keep test files for future work.

Also updates the charter.md to indicate we've moved TSC information into
the docs folder in the community/TSC.rst file.

Signed-off-by: David B. Kinder <david.b.kinder@intel.com>